+++ ResNet.WebAPI/Controllers/TableController.cs
@@ -1,3 +1,4 @@
+using System.Net;
 using Domain.Constants;
 using Domain.Responses;
 using Infrastructure.Interfaces;
@@ -10,7 +11,7 @@ namespace WebApi.Controllers;
 
 [ApiController]
 [Route("api/restaurants/{restaurantId:int}/tables")]
-public class TableController(ITableService tableService)
+public class TableController(ITableService tableService) : ControllerBase
 {
     [HttpGet]
     public async Task<PagedResponse<List<GetTableDto>>> GetTablesByRestaurantId(int restaurantId, [FromQuery] TableFilter filter)
@@ -25,25 +26,34 @@ public class TableController(ITableService tableService)
     }
 
     [HttpPost]
-    [Authorize(Roles = Roles.Admin)]
+    [Authorize(Roles = Roles.Admin + "," + Roles.Owner)]
     public async Task<Response<GetTableDto>> AddTable(int restaurantId, CreateTableDto tableDto)
     {
+        if (!await HasAccessToRestaurant(restaurantId))
+            return new Response<GetTableDto>(HttpStatusCode.Forbidden, "Access denied");
+
         tableDto.RestaurantId = restaurantId;
         return await tableService.AddTableAsync(tableDto);
     }
 
     [HttpPut("{id:int}")]
-    [Authorize(Roles = Roles.Admin)]
+    [Authorize(Roles = Roles.Admin + "," + Roles.Owner)]
     public async Task<Response<GetTableDto>> UpdateTable(int restaurantId, int id, UpdateTableDto tableDto)
     {
+        if (!await HasAccessToRestaurant(restaurantId))
+            return new Response<GetTableDto>(HttpStatusCode.Forbidden, "Access denied");
+
         tableDto.RestaurantId = restaurantId;
         return await tableService.UpdateTableAsync(id, tableDto);
     }
 
     [HttpDelete("{id:int}")]
-    [Authorize(Roles = Roles.Admin)]
+    [Authorize(Roles = Roles.Admin + "," + Roles.Owner)]
     public async Task<Response<string>> DeleteTable(int restaurantId, int id)
     {
+        if (!await HasAccessToRestaurant(restaurantId))
+            return new Response<string>(HttpStatusCode.Forbidden, "Access denied");
+
         return await tableService.DeleteTableAsync(id);
     }
 
@@ -58,5 +68,17 @@ public class TableController(ITableService tableService)
     public async Task<Response<List<GetTableDto>>> GetAvailableTables(int restaurantId, [FromQuery] DateTime? dateTime, [FromQuery] TimeSpan? duration)
     {
         return await tableService.GetAvailableTablesAsync(restaurantId, dateTime, duration);
+    }
+
+    private async Task<bool> HasAccessToRestaurant(int restaurantId)
+    {
+        if (User.IsInRole(Roles.Admin))
+            return true;
+
+        var claim = User.FindFirst("restaurant_id")?.Value;
+        if (string.IsNullOrEmpty(claim))
+            return false;
+
+        return int.Parse(claim) == restaurantId;
     }
 }
